+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-# dbt_monitorial_snowflake_cortex Changelog
+# dbt-snowflake-cortex Changelog
 
-## 1.0.0 - Snowflake Agents
+## 1.0.0 - 2026-03-17 - Snowflake Agents
 
 * added in new materalization `agent` which allows you to create and manage Snowflake Cortex Agents
 * added in macros for grants of `semantic views` and `agents`